[Chore] #43 - 설문조사 수정
- 필요없는 api 서비스 삭제
- 설문조사 req int에서 float로 변경
- 설문조사에서 Post한 후 받는 Res는 Unit으로 변경.(Res 없음)
- ViewModel도 수정

+++ app/src/main/java/com/example/fit4you_android/data/dto/request/BaseQuestionReq.kt
@@ -2,11 +2,11 @@ package com.example.fit4you_android.data.dto.request
 
 data class BaseQuestionReq(
     var email: String,
-    var neck: Int,
-    var shoulder: Int,
-    var lumbar: Int,
-    var wrist: Int,
-    var elbow: Int,
-    var knee: Int,
+    var neck: Float,
+    var shoulder: Float,
+    var lumbar: Float,
+    var wrist: Float,
+    var elbow: Float,
+    var knee: Float,
     var hist: List<String>
 )
